[REFACTOR]: dependencies version status code review & refactor
