+++ v6y-apps/bfb-static-auditor/src/auditors/dependencies/DependenciesAuditor.ts
