+++ v6y-apps/bfb-static-auditor/src/auditors/dependencies/DependenciesUtils.ts
