+++ v6y-apps/bfb-static-auditor/src/auditors/dependencies/DependencyVulnerabilityAnalyzer.ts
